+++ src/Dofus/Messages/AuthenticationTicketAcceptedMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class AuthenticationTicketAcceptedMessage : INetworkMessage
     {
-        internal static int MessageId => 8662;
+        internal static ushort MessageId => 8662;
 
         public void Serialize(DofusBinaryWriter writer)
         {
