+++ src/Dofus/Messages/AuthenticationTicketRefusedMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class AuthenticationTicketRefusedMessage : INetworkMessage
     {
-        internal static int MessageId => 3932;
+        internal static ushort MessageId => 3932;
 
         public void Serialize(DofusBinaryWriter writer)
         {
