+++ src/Dofus/Messages/CredentialsAcknowledgementMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class CredentialsAcknowledgementMessage : INetworkMessage
     {
-        internal static int MessageId => 5417;
+        internal static ushort MessageId => 5417;
 
         public void Serialize(DofusBinaryWriter writer)
         {
