+++ src/Dofus/Messages/HelloGameMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class HelloGameMessage : INetworkMessage
     {
-        internal static int MessageId => 5556;
+        internal static ushort MessageId => 5556;
 
         public void Serialize(DofusBinaryWriter writer)
         {
