+++ src/Dofus/Messages/LoginQueueStatusMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class LoginQueueStatusMessage : INetworkMessage
     {
-        internal static int MessageId => 2553;
+        internal static ushort MessageId => 2553;
         public ushort Position { get; private set; }
         public ushort Total { get; private set; }
