+++ src/Dofus/Messages/PrismListUpdateMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class PrismListUpdateMessage : INetworkMessage
     {
-        internal static int MessageId => 101;
+        internal static ushort MessageId => 101;
 
         public void Serialize(DofusBinaryWriter writer)
         {
