+++ src/Dofus/Messages/ProtocolRequiredMessage.cs
@@ -5,7 +5,7 @@ namespace Dofus.Messages
 {
     public class ProtocolRequiredMessage : INetworkMessage
     {
-        internal static int MessageId => 9546;
+        internal static ushort MessageId => 9546;
         public string Version { get; private set; } = string.Empty;
 
         public void Serialize(DofusBinaryWriter writer) => throw new NotImplementedException();
