+++ src/Dofus/Messages/SequenceNumberRequestMessage.cs
@@ -4,7 +4,7 @@ namespace Dofus.Messages
 {
     public class SequenceNumberRequestMessage : INetworkMessage
     {
-        internal static int MessageId => 3495;
+        internal static ushort MessageId => 3495;
 
         public void Serialize(DofusBinaryWriter writer)
         {
