+++ src/Dofus/Types/ActorOrientation.cs
@@ -5,7 +5,7 @@ namespace Dofus.Types
 {
     public class ActorOrientation : INetworkMessage
     {
-        internal static int MessageId => 6675;
+        internal static ushort MessageId => 6675;
 
         public long Id { get; private set; }
         public byte Direction { get; private set; }
